Enable fractional key_stroke_count for fine-grained typing metrics
Change key_stroke_count from int to float to support fractional
key stroke decreases from the auto-decrease mechanism. The API
boundary still sends int(keystroke_count) for server backward
compatibility. The internal display format uses ".2f" to show
two decimal places.

Constraint: API boundary sends int for server backward compatibility
Rejected: Keep as int and round to nearest integer | loses precision needed for fine-grained decrease
Rejected: Add a separate fractional field | unnecessary complexity, simpler to change the type directly
Confidence: high
Scope-risk: narrow
Tested: 439 pytest pass, ruff check pass, ruff format pass

+++ src/backend/models/entity/session_stat.py
@@ -21,7 +21,7 @@ class SessionStat:
 
     # 打字统计数据
     time: float = 0.0
-    key_stroke_count: int = 0
+    key_stroke_count: float = 0.0
     char_count: int = 0
     wrong_char_count: int = 0
     backspace_count: int = 0
